// src/model/useProductModel.tsx
export type NewCollectionList = NewCollectionListData[];

export interface NewCollectionListData {
  activation_date?: string;
  address: string;
  ask_for_price: string;
  brand: string;
  brochure: string;
  category: string;
  city: string;
  condition: string;
  country: string;
  create_time: string;
  created_by: string;
  dealstatus?: string;
  description: string;
  disable_email_contact: string;
  email: string;
  expirtion_date?: string;
  featured: string;
  featured_expiration_date: any;
  featured_img: string;
  featured_img_thumb: string;
  fuel_type: string;
  gallery: string;
  hide_email: string;
  hide_phone: string;
  id: string;
  is_verified?: string;
  last_update_time: string;
  latitude: string;
  longitude: string;
  mileage: string;
  model: string;
  office_address: string;
  office_name: string;
  parent_category: string;
  phone_no: string;
  price: string;
  publish_time: string;
  purpose: string;
  rating: string;
  report: string;
  search_meta: string;
  show_office?: string;
  state: string;
  status: string;
  tags: string;
  title: string;
  total_view: string;
  transmission: string;
  unique_id: string;
  video_url: string;
  year: string;
  zip_code: string;
}